Move JaspiConfiguration to jetty-webapp so it works with JPMS.
