--- jetty-jaspi/src/main/resources/META-INF/services/org.eclipse.jetty.webapp.Configuration
@@ -1,1 +1,0 @@
-org.eclipse.jetty.security.jaspi.JaspiConfiguration
