+++ jetty-webapp/src/main/resources/META-INF/services/org.eclipse.jetty.webapp.Configuration
@@ -1,6 +1,7 @@
 org.eclipse.jetty.webapp.FragmentConfiguration
 org.eclipse.jetty.webapp.JettyWebXmlConfiguration
 org.eclipse.jetty.webapp.JaasConfiguration
+org.eclipse.jetty.webapp.JaspiConfiguration
 org.eclipse.jetty.webapp.JmxConfiguration
 org.eclipse.jetty.webapp.JndiConfiguration
 org.eclipse.jetty.webapp.JspConfiguration
